+++ src/EngineBuilder.php
@@ -5,8 +5,6 @@ declare(strict_types = 1);
 namespace hanneskod\readmetester;
 
 use hanneskod\readmetester\Example\ExampleFactory;
-use hanneskod\readmetester\Example\ProcessorInterface;
-use hanneskod\readmetester\Example\ProcessorContainer;
 use hanneskod\readmetester\Parser\Parser;
 use hanneskod\readmetester\Expectation\ExpectationEvaluator;
 use hanneskod\readmetester\Expectation\ExpectationFactory;
@@ -14,19 +12,7 @@ use hanneskod\readmetester\Runner\RunnerInterface;
 
 class EngineBuilder
 {
-    private ProcessorInterface $processor;
     private RunnerInterface $runner;
-
-    public function __construct()
-    {
-        $this->processor = new ProcessorContainer;
-    }
-
-    public function setProcessor(ProcessorInterface $processor): self
-    {
-        $this->processor = $processor;
-        return $this;
-    }
 
     public function setRunner(RunnerInterface $runner): self
     {
@@ -40,7 +26,6 @@ class EngineBuilder
             new Parser,
             new ExampleFactory(
                 new ExpectationFactory,
-                $this->processor
             ),
             new ExampleTester(
                 $this->getRunner(),