+++ src/Example/ProcessorInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace hanneskod\readmetester\Example;
-
-// TODO replace with TransformationInterface
-interface ProcessorInterface
-{
-    public function process(ExampleObj $example): ExampleObj;
-}